+++ src/Products/PloneMeeting/tests/testAnnexes.py
@@ -22,8 +22,8 @@ from imio.annex.columns import ActionsColumn
 from imio.annex.utils import get_annexes_to_print
 from imio.helpers.content import get_vocab
 from imio.helpers.content import get_vocab_values
+from imio.helpers.content import richtextval
 from plone import api
-from plone.app.textfield.value import RichTextValue
 from plone.dexterity.utils import createContentInContainer
 from plone.indexer.wrapper import IndexableObjectWrapper
 from Products.Archetypes.event import ObjectEditedEvent
@@ -361,7 +361,7 @@ class testAnnexes(PloneMeetingTestCase):
             'meetingadvice',
             **{'advice_group': self.vendors_uid,
                'advice_type': u'positive',
-               'advice_comment': RichTextValue(u'My comment')})
+               'advice_comment': richtextval(u'My comment')})
         # enable confidentiality
         self._enable_annex_config(advice)
         annexes_table = advice.restrictedTraverse('@@iconifiedcategory')
@@ -1708,7 +1708,7 @@ class testAnnexes(PloneMeetingTestCase):
             'meetingadvice',
             **{'advice_group': self.developers_uid,
                'advice_type': u'positive',
-               'advice_comment': RichTextValue(u'My comment')})
+               'advice_comment': richtextval(u'My comment')})
         _check_catalog()
         advice_annex = self.addAnnex(advice)
         _check_catalog()